Final CSS touches and guest cart refactor

+++ client/store/cart.js
@@ -14,32 +14,71 @@ const deletedMovie = movieId => ({
   type: DELETE_MOVIE,
   movieId
 })
-
-//Thunk
-export const addToCartThunk = (userId, movieId) => async dispatch => {
-  try {
-    const {data} = await axios.post(`/api/users/${userId}/cart`, {movieId})
-  } catch (err) {
-    console.error(err)
+export const getCartThunk = userId => async dispatch => {
+  if (userId !== undefined) {
+    try {
+      console.log('before axios')
+      const {data} = await axios.get(`/api/users/${userId}/cart`)
+      console.log('the data', data)
+      dispatch(getCart(data))
+    } catch (err) {
+      console.error(err)
+    }
+  } else {
+    try {
+      const data = JSON.parse(localStorage.getItem('items'))
+      dispatch(getCart(data))
+    } catch (err) {
+      console.error(err)
+    }
   }
 }
 
-export const getCartThunk = userId => async dispatch => {
-  try {
-    const {data} = await axios.get(`/api/users/${userId}/cart`)
-    dispatch(getCart(data))
-  } catch (err) {
-    console.error(err)
+//Thunk
+export const addToCartThunk = (user, movieId, movie) => async dispatch => {
+  if (user) {
+    try {
+      await axios.post(`/api/users/${user.id}/cart`, {movieId})
+      const {data} = await axios.get(`/api/users/${user.id}/cart`)
+      dispatch(getCart(data))
+    } catch (err) {
+      console.error(err)
+    }
+  } else {
+    try {
+      console.log('here!')
+      let itemsArray = localStorage.getItem('items')
+        ? JSON.parse(localStorage.getItem('items'))
+        : []
+
+      console.log(movie)
+      itemsArray.push(movie)
+      localStorage.setItem('items', JSON.stringify(itemsArray))
+      dispatch(getCart(itemsArray))
+    } catch (err) {
+      console.error(err)
+    }
   }
 }
 
 export const deleteMovieFromCart = (userId, movieId) => {
   return async dispatch => {
-    try {
-      await axios.delete(`/api/users/${userId}/cart/${movieId}`)
-      dispatch(deletedMovie(movieId))
-    } catch (error) {
-      console.error(error)
+    if (!userId) {
+      try {
+        const data = JSON.parse(localStorage.getItem('items'))
+        data.splice(movieId, 1)
+        localStorage.setItem('items', JSON.stringify(data))
+        dispatch(getCart(data))
+      } catch (err) {
+        console.error(err)
+      }
+    } else {
+      try {
+        await axios.delete(`/api/users/${userId}/cart/${movieId}`)
+        dispatch(deletedMovie(movieId))
+      } catch (error) {
+        console.error(error)
+      }
     }
   }
 }
